Rename dummyserver.server to dummyserver.tornadoserver
