+++ test/contrib/test_socks.py
@@ -11,8 +11,8 @@ from unittest.mock import Mock, patch
 import pytest
 import socks as py_socks  # type: ignore[import]
 
-from dummyserver.server import DEFAULT_CA, DEFAULT_CERTS
 from dummyserver.testcase import IPV4SocketDummyServerTestCase
+from dummyserver.tornadoserver import DEFAULT_CA, DEFAULT_CERTS
 from urllib3.contrib import socks
 from urllib3.exceptions import ConnectTimeoutError, NewConnectionError
