+++ test/with_dummyserver/test_socketlevel.py
@@ -24,13 +24,13 @@ from unittest import mock
 import pytest
 import trustme
 
-from dummyserver.server import (
+from dummyserver.testcase import SocketDummyServerTestCase, consume_socket
+from dummyserver.tornadoserver import (
     DEFAULT_CA,
     DEFAULT_CERTS,
     encrypt_key_pem,
     get_unreachable_address,
 )
-from dummyserver.testcase import SocketDummyServerTestCase, consume_socket
 from urllib3 import HTTPConnectionPool, HTTPSConnectionPool, ProxyManager, util
 from urllib3._collections import HTTPHeaderDict
 from urllib3.connection import HTTPConnection, _get_default_user_agent
